Ignore advice in CombinedExtractor when forwarding (not __init__)
Save/load messes with __init__ and causes problems when an actor is
re-loaded as an advisor (the actor saw advice, not the advisor). This
commit fully ignores advice in a way that is opaque to the network
(advice is always zeroed-out in the network input, so does not influence
the output but is "here").